pyproject.toml
[project]
name = "rag3"
version = "0.1.0"
description = "Production RAG System built on Haystack"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "haystack-ai>=2.12.0",
    "ollama-haystack>=2.2.0",
    "pgvector-haystack>=1.1.0",
    "unstructured[pdf]>=0.15.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0",
    "psycopg2-binary>=2.9.11",
    "tenacity>=8.2.0",
    "tiktoken>=0.8.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
eval = [
    "ragas>=0.1.0",
    "datasets>=2.14.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
